Merging Development -> Master.

+++ FoodDeliveryApp/Controllers/ProductController.cs
@@ -1,8 +1,11 @@
 ﻿using FoodDeliveryApp.Core.Contracts;
 using FoodDeliveryApp.Core.Models.Product;
+using FoodDeliveryApp.Infrastructure.Data.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using static FoodDeliveryApp.Core.Constants.ErrorMessagesConstants.ProductErrorMessagesConstants;
+using static FoodDeliveryApp.Core.Constants.ErrorMessagesConstants.UserErrorMessagesConstants;
 using static FoodDeliveryApp.Core.Constants.MessageConstants.ProductMessageConstants;
 
 namespace FoodDeliveryApp.Controllers
@@ -37,7 +40,12 @@ namespace FoodDeliveryApp.Controllers
 		[HttpGet]
 		public async Task<IActionResult> Add()
 		{
-			var model = new ProductFormModel()
+            if (User.IsAdmin() == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = NotAdminErrorMessage });
+            }
+
+            var model = new ProductFormModel()
 			{
 				Categories = await productService.GetCategoriesAsync(),
 				SpicyCategories = await productService.GetSpicyCategoriesAsync()
@@ -49,7 +57,12 @@ namespace FoodDeliveryApp.Controllers
 		[HttpPost]
 		public async Task<IActionResult> Add(ProductFormModel model, int restaurantId)
 		{
-			if (await restaurantService.ExistsRestaurantAsync(restaurantId) == false)
+            if (User.IsAdmin() == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = NotAdminErrorMessage });
+            }
+
+            if (await restaurantService.ExistsRestaurantAsync(restaurantId) == false)
 			{
 				ModelState.AddModelError(nameof(restaurantId), InvalidRestaurantMessage);
 			}
@@ -82,12 +95,19 @@ namespace FoodDeliveryApp.Controllers
 		[HttpGet]
 		public async Task<IActionResult> Edit(int productId)
 		{
-			if (await productService.ExistsProductAsync(productId) == false)
+            if (User.IsAdmin() == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = NotAdminErrorMessage });
+            }
+
+            if (await productService.ExistsProductAsync(productId) == false)
 			{
 				return RedirectToAction("Error", "Home", new { errorMessage = InvalidProductErrorMessage });
 			}
 
 			var model = await productService.GetProductFormModelByIdAsync(productId);
+
+			ViewBag.ProductId = productId;
 
 			return View(model);
 		}
@@ -95,7 +115,12 @@ namespace FoodDeliveryApp.Controllers
 		[HttpPost]
 		public async Task<IActionResult> Edit(ProductFormModel model, int productId)
 		{
-			if (await productService.ExistsProductAsync(productId) == false)
+            if (User.IsAdmin() == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = NotAdminErrorMessage });
+            }
+
+            if (await productService.ExistsProductAsync(productId) == false)
 			{
 				return RedirectToAction("Error", "Home", new { errorMessage = InvalidProductErrorMessage });
 			}
@@ -118,20 +143,44 @@ namespace FoodDeliveryApp.Controllers
 			return RedirectToAction(nameof(Details), new { productId });
 		}
 
-		[HttpPost]
-		public async Task<IActionResult> Delete(int productId)
+        [HttpGet]
+        public async Task<IActionResult> Delete(int productId)
+        {
+            if (User.IsAdmin() == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = NotAdminErrorMessage });
+            }
+
+            if (await productService.ExistsProductAsync(productId) == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = InvalidProductErrorMessage });
+            }
+
+            var model = await productService.GetProductDeleteModelByIdAsync(productId);
+
+            return View(model);
+        }
+
+        [HttpPost]
+		public async Task<IActionResult> DeleteConfirmed(int productId)
 		{
-			var product = await productService.GetProductByIdAsync(productId);
+            if (User.IsAdmin() == false)
+            {
+                return RedirectToAction("Error", "Home", new { errorMessage = NotAdminErrorMessage });
+            }
+
+            var product = await productService.GetProductByIdAsync(productId);
 
 			if (product == null)
 			{
 				return RedirectToAction("Error", "Home", new { errorMessage = InvalidProductErrorMessage });
-			
 			}
+
+			int restaurantId = product.RestaurantId;
 
 			await productService.DeleteProductAsync(productId);
 
-			return RedirectToAction("Menu", "Restaurant", new { product.RestaurantId });
+			return RedirectToAction("Menu", "Restaurant", new { restaurantId });
 		}
 	}
 }
